==> preprocess/30-colab-structured_description.ipynb <==
# 사고 설명문 구조화 with LLM

### colab 세팅

from google.colab import drive
drive.mount('/content/drive')

parent_path = '/content/drive/Othercomputers/My MacBook Pro/colab/'

import sys
sys.path.append(parent_path + 'preprocess')

!python '/content/drive/Othercomputers/My MacBook Pro/colab/settings.py'
%env OLLAMA_MAX_LOADED_MODELS=2
%env OLLAMA_NUM_PARALLEL=2
!nohup ollama serve &

### 구조화 요청 prompt 준비

import pandas as pd
from utils import load_ollama, run_ollama, process_dataframe_in_chunks
import random

test_df = pd.read_csv(parent_path + 'data/test_02.csv')
train_df = pd.read_csv(parent_path + 'data/train_02.csv')

def make_structuring_prompt(row):
    prompt = f"""한국 건설 공사 안전 사고 관련 문장을 언어 모델에 사용하기 위해 전처리하려 합니다.
사고가 발생한 배경과 핵심적인 피해 내용, 사고 원인을 도출하고, \"발생 배경, 사고 종류, 사고 원인\"을 json 형식으로 핵심적인 내용만 간결히 정리해주세요.
추가적인 정보를 임의로 추론하거나 생성하지 말고, 원문에 주어진 정보만을 반영하세요.
제공될 문장은 [문장: \"사고 종류 (사고 설명 또는 원인)\"] 형식으로 제공됩니다.
문장: \"{row['correct_description']}\""""
    return prompt
    # '피해 내용' 삭제
def model_query_test(df, prompt_func, chain, idx=-1):
    if idx == -1:
        random_idx = random.randint(0, len(df) - 1)
    else:
        random_idx = idx
    print(f"idx: {random_idx}")

    prompt = prompt_func(df.iloc[random_idx])
    print(prompt)

    if df.loc[random_idx, 'category_exists'] == "N":
        print("사고 종류 없으므로 그대로 사용.")
        return

    answer = run_ollama(prompt, chain)
    print(answer)

chain = load_ollama(model="gemma3:4b", temperature=0.0)

train_df['structuring_prompt'] = train_df.apply(lambda x: make_structuring_prompt(x), axis=1)
test_df['structuring_prompt'] = test_df.apply(lambda x: make_structuring_prompt(x), axis=1)

### 쿼리 테스트

model_query_test(train_df, make_structuring_prompt, chain, idx=-1) #21459

### 모델 준비 및 실행

# 모델 준비
chain1 = load_ollama(model="gemma3:4b", temperature=0.0)
chain2 = load_ollama(model="gemma3:4bcp", temperature=0.0)

no_category_test_df = test_df[test_df['category_exists']=="Y"]
no_category_train_df = train_df[train_df['category_exists']=="Y"]

process_dataframe_in_chunks(no_category_test_df, 'structuring_prompt', 'structured_json', chain1, chain2)
test_df.loc[test_df['category_exists']=="Y", 'structured_json'] = no_category_test_df['structured_json']
test_df.to_csv(parent_path + 'data/test_03.csv', index=False)

!nohup ollama serve &

process_dataframe_in_chunks(no_category_train_df, 'structuring_prompt', 'structured_json', chain1, chain2, chunk_size=30)
train_df.loc[train_df['category_exists']=="Y", 'structured_json'] = no_category_train_df['structured_json']
train_df.to_csv(parent_path + 'data/train_03.csv', index=False)
